+++ client/controllers/mirrorcontroller.hpp
@@ -113,7 +113,7 @@ private:
     UaStatus saveMPESPositions(const std::string &saveFilePath);
     UaStatus savePanelTemperatures(const std::string &saveFilePath);
 
-    std::vector<float> getAzEl();
+    std::vector<double> getAzEl();
 
     // Align all edges fron need_alignment starting at start_idx and  moving in the direction dir
     UaStatus alignSequential(const std::string &startEdge, const std::string &EndEdge, unsigned dir);
